backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.1
openai==1.6.1
requests==2.27.1
beautifulsoup4==4.10.0
lxml==4.9.3
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
supabase==2.0.3
pytest==7.4.0
aiohttp==3.8.5
gunicorn==21.2.0
youtube-transcript-api==0.6.1 